change `ModelRelatedField` to be a generic interface
